+++ db/seed_db.js
@@ -1,3 +1,4 @@
+const { name } = require("ejs");
 const prisma = require("./prisma");
 const bcrypt = require("bcryptjs");
 
@@ -37,58 +38,37 @@ const accounts = Promise.all(
   .catch((err) => console.log(err.message));
 
 const seedDB = async () => {
+  await prisma.file.deleteMany();
   await prisma.folder.deleteMany();
   await prisma.account.deleteMany();
-  await prisma.file.deleteMany();
-  // await prisma.account.createMany({
-  //   data: await accounts,
-  // });
 
   const account = await prisma.account.create({
     data: {
-      name: "Bill Dauterive",
-      username: "bill_dozer",
-      email: "bill.d@gmail.com",
-      password: "Test123!",
-      drive: {
+      name: "Luanne Platter",
+      username: "platter_lp",
+      email: "lp_barber@gmail.com",
+      password: "manGer1&BabY",
+      folders: {
         create: {
-          name: "Root folder",
-          subFolders: {
-            create: {
-              name: "A nested folder",
-              files: {
-                create: {
-                  name: "File in nested folder",
-                },
-              },
-              subFolders: {
-                create: {
-                  name: "another nested folder",
-                },
-              },
-            },
+          name: "Folder 1",
+          files: {
+            create: {},
           },
         },
       },
     },
     include: {
-      drive: {
+      folders: {
         include: {
-          files: true,
-          subFolders: {
-            include: {
-              files: true,
-              subFolders: true,
-            },
-          },
+          files: {},
         },
       },
     },
   });
 
   console.log("account:", account);
-  // console.log("account.drive.files:", account.drive[0].files);
-  // console.log("account.drive.subFolders:", account.drive[0].subFolders);
+  console.log("account.folders:", account.folders);
+  console.log("account.folders[0].files:", account.folders[0].files);
 };
 
 seedDB();