+++ prisma/migrations/20250212165311_testing_user_to_many/migration.sql
@@ -1,0 +1,5 @@
+-- DropIndex
+DROP INDEX "Directory_parentDirId_key";
+
+-- DropIndex
+DROP INDEX "Directory_userId_key";