+++ prisma/migrations/20250212171727_testing_rm_directory_id/migration.sql
@@ -1,0 +1,11 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `directoryId` on the `User` table. All the data in the column will be lost.
+
+*/
+-- DropIndex
+DROP INDEX "User_directoryId_key";
+
+-- AlterTable
+ALTER TABLE "User" DROP COLUMN "directoryId";