+++ prisma/migrations/20250212183246_account_folders/migration.sql
@@ -1,0 +1,22 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `parentFolderId` on the `Folder` table. All the data in the column will be lost.
+  - Made the column `accountId` on table `Folder` required. This step will fail if there are existing NULL values in that column.
+
+*/
+-- DropForeignKey
+ALTER TABLE "Folder" DROP CONSTRAINT "Folder_accountId_fkey";
+
+-- DropForeignKey
+ALTER TABLE "Folder" DROP CONSTRAINT "Folder_parentFolderId_fkey";
+
+-- DropIndex
+DROP INDEX "Folder_parentFolderId_key";
+
+-- AlterTable
+ALTER TABLE "Folder" DROP COLUMN "parentFolderId",
+ALTER COLUMN "accountId" SET NOT NULL;
+
+-- AddForeignKey
+ALTER TABLE "Folder" ADD CONSTRAINT "Folder_accountId_fkey" FOREIGN KEY ("accountId") REFERENCES "Account"("id") ON DELETE RESTRICT ON UPDATE CASCADE;