+++ prisma/migrations/20250212231827_account_id_required/migration.sql
@@ -1,0 +1,2 @@
+-- DropForeignKey
+ALTER TABLE "File" DROP CONSTRAINT "File_accountId_fkey";