+++ prisma/migrations/20250213031159_account_files/migration.sql
@@ -1,0 +1,2 @@
+-- AddForeignKey
+ALTER TABLE "File" ADD CONSTRAINT "File_accountId_fkey" FOREIGN KEY ("accountId") REFERENCES "Account"("id") ON DELETE RESTRICT ON UPDATE CASCADE;