+++ prisma/migrations/20250213041710_testing_files_folders/migration.sql
@@ -1,0 +1,14 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `folderId` on the `File` table. All the data in the column will be lost.
+
+*/
+-- DropForeignKey
+ALTER TABLE "File" DROP CONSTRAINT "File_folderId_fkey";
+
+-- AlterTable
+ALTER TABLE "File" DROP COLUMN "folderId";
+
+-- AddForeignKey
+ALTER TABLE "File" ADD CONSTRAINT "File_accountId_fkey" FOREIGN KEY ("accountId") REFERENCES "Account"("id") ON DELETE RESTRICT ON UPDATE CASCADE;